+++ src/main/java/ru/maximkulikov/goodgame/api/chatmodels/ResChannelCounters.java
@@ -16,9 +16,9 @@ public class ResChannelCounters implements ResChatObject {
     private String channelId;
 
     @JsonProperty("clients_in_channel")
-    private Long clientsInChannel;
+    private Integer clientsInChannel;
 
     @JsonProperty("users_in_channel")
-    private Long usersInChannel;
+    private Integer usersInChannel;
 
 }
