+++ src/main/java/ru/maximkulikov/goodgame/api/chatmodels/ResUsersList.java
@@ -17,10 +17,10 @@ public class ResUsersList implements ResChatObject {
     private String channelId;
 
     @JsonProperty("clients_in_channel")
-    private Long clientsInChannel;
+    private Integer clientsInChannel;
 
     @JsonProperty("users_in_channel")
-    private Long usersInChannel;
+    private Integer usersInChannel;
 
     private List<ChatUser> users;
